refactor: Split wmts-playground into files
