+++ app/charts/map/types.ts
@@ -1,0 +1,4 @@
+import { ParsedWMSLayer } from "./wms-utils";
+import { ParsedWMTSLayer } from "./wmts-utils";
+
+export type CustomLayer = ParsedWMSLayer | ParsedWMTSLayer;
